delete, update user validation

+++ src/main/java/com/studymate/web/QuestionController.java
@@ -11,6 +11,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.studymate.domain.Question;
 import com.studymate.domain.QuestionRepository;
@@ -25,13 +26,13 @@ public class QuestionController {
 	
 	@GetMapping("/form")
 	public String form(HttpSession session) {
-		if(!HttpSessionUtils.isLoginUser(session)) return "/users/signIn";
+		if(!HttpSessionUtils.isLoginUser(session)) return "redirect:/users/signIn";
 		return "/qna/form";
 	}
 	
 	@PostMapping("")
 	public String create(String title, String contents, HttpSession session) {
-		if(!HttpSessionUtils.isLoginUser(session)) return "/users/signIn";
+		if(!HttpSessionUtils.isLoginUser(session)) return "redirect:/users/signIn";
 		
 		User sessionUser = HttpSessionUtils.getUserFromSession(session);
 	
@@ -49,9 +50,15 @@ public class QuestionController {
 	}
 	
 	@GetMapping("/{id}/form")
-	public String questionUpdateForm(@PathVariable Long id, Model model) {
+	public String questionUpdateForm(@PathVariable Long id, Model model, HttpSession session) {
+		if(!HttpSessionUtils.isLoginUser(session)) return "redirect:/users/signIn";
+		
 		Question question = quesitonRepository.findById(id).orElse(null);
-		if(question==null) return "redirect:/questions/{id}";
+		if(question==null) return "redirect:/";
+		
+		User sessionUser = HttpSessionUtils.getUserFromSession(session);
+		
+		if(!question.isSameWriter(sessionUser)) return String.format("redirect:/questions/%d", id);
 		model.addAttribute("question", question);
 		return "/qna/updateForm";
 	}
@@ -60,14 +67,19 @@ public class QuestionController {
 	public String updateQuestion(@PathVariable Long id, String title, String contents) {
 		Question question = quesitonRepository.findById(id).orElse(null);
 		if(question==null) return "redirect:/";
-		question.update(title, contents);
+		question.update(title, contents);	
 		quesitonRepository.save(question);
 		return String.format("redirect:/questions/%d", id);
 	}
 	
 	@DeleteMapping("/{id}")
-	public String delete(@PathVariable Long id) {
+	public String delete(@PathVariable Long id, HttpSession session) {
+		if(!HttpSessionUtils.isLoginUser(session)) return "redirect:/users/signIn";
+		
 		Question question = quesitonRepository.findById(id).orElse(null);
+		User sessionUser = HttpSessionUtils.getUserFromSession(session);
+		if(!question.isSameWriter(sessionUser)) return String.format("redirect:/questions/%d", id);
+		
 		quesitonRepository.delete(question);
 		return "redirect:/";
 	}
